% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_utils.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean database NOAA}
\source{
This capstone project will be centered around a dataset obtained from the U.S. National Oceanographic
and Atmospheric Administration (NOAA) on significant earthquakes around the world. This dataset contains
information about 5,933 earthquakes over an approximately 4,000 year time span.
}
\usage{
eq_clean_data(earthquakes)
}
\arguments{
\item{earthquakes}{to enter a database of NOAA.}
}
\value{
if file exists, this function read the file and return a database as a data frame where date, latitude and longitude
have special formats.
}
\description{
The function \code{eq_clean_data} read a database and returns a database with special formats of some columns.
}
\examples{
\dontrun{eq_clean_data("earthquakes.tsv.gz")}

}
\seealso{
\code{\link{eq_location_clean}}
}
