% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_utils.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Format to location in the database}
\source{
This capstone project will be centered around a dataset obtained from the U.S. National Oceanographic
and Atmospheric Administration (NOAA) on significant earthquakes around the world. This dataset contains
information about 5,933 earthquakes over an approximately 4,000 year time span.
}
\usage{
eq_location_clean(location)
}
\arguments{
\item{location}{charater string with location and country followed by a colon.}
}
\value{
This function returns a character string in title case with the country and colon removed.
}
\description{
The function \code{eq_location_clean} clean earthquake location data.
}
\examples{
\dontrun{eq_location_clean("JAPAN:  TUWANO")}

}
